Refactor background script
