+++ packages/light-apps/package.json
@@ -21,7 +21,7 @@
   "dependencies": {
     "@polkadot/api": "^1.3.1",
     "@polkadot/ui-assets": "^0.51.1",
-    "@substrate/context": "^0.3.20",
+    "@substrate/context": "^0.3.21",
     "@substrate/ui-components": "^0.3.0",
     "@types/chrome": "^0.0.95",
     "@types/file-saver": "^2.0.0",
