+++ packages/ui-components/package.json
@@ -33,11 +33,11 @@
     "@storybook/addon-knobs": "^5.2.5",
     "@storybook/addon-links": "^5.2.5",
     "@storybook/react": "^5.2.5",
-    "@substrate/context": "^0.3.20",
     "@types/react-color": "^3.0.1",
     "@types/webpack-env": "^1.14.0"
   },
   "peerDependencies": {
+    "@substrate/context": "^0.3.21",
     "react": "^16.9"
   }
 }
